src/lambdas/requirements.txt
pandas
mail-parser
openpyxl
xlrd
python-slugify
awswrangler
requests
